Fix error with output dir

+++ setup.py
@@ -5,15 +5,16 @@ def setup_output_dir(output_dir: str):
 
     # Removing all the contents of output_dir:
 
-    for obj in os.listdir(output_dir):
+    if os.path.isdir(output_dir):
+        for obj in os.listdir(output_dir):
 
-        path = os.path.join(output_dir, obj)
+            path = os.path.join(output_dir, obj)
 
-        if os.path.isfile(path) or os.path.islink(path):
-            os.unlink(path)
+            if os.path.isfile(path) or os.path.islink(path):
+                os.unlink(path)
 
-        elif os.path.isdir(path):
-            shutil.rmtree(path)
+            elif os.path.isdir(path):
+                shutil.rmtree(path)
